+++ assets/javascripts/kitten/components/cards/reward-card.js
@@ -44,6 +44,9 @@ class RewardCardComponent extends Component {
     valueAvailability: PropTypes.string,
 
     button: PropTypes.string,
+    buttonOnMouseEnter: PropTypes.func,
+    buttonOnMouseLeave: PropTypes.func,
+    buttonOnClick: PropTypes.func,
 
     myContribution: PropTypes.string,
     manageContribution: PropTypes.string,
@@ -64,12 +67,40 @@ class RewardCardComponent extends Component {
   }
 
   render() {
-    const { isDisabled, viewportIsSOrLess } = this.props
+    const {
+      isDisabled,
+      viewportIsSOrLess,
+      viewportIsMobile,
+      viewportIsTabletOrLess,
+      titleAmount,
+      titleDescription,
+      textDescription,
+      titleContributors,
+      titleSmallContributors,
+      titleDelivery,
+      titleSmallDelivery,
+      titleAvailability,
+      titleSmallAvailability,
+      valueContributors,
+      valueDelivery,
+      valueAvailability,
+      button,
+      buttonOnMouseEnter,
+      buttonOnMouseLeave,
+      buttonOnClick,
+      myContribution,
+      manageContribution,
+      manageContributionLink,
+      imageSrc,
+      imageSrcSmall,
+      imageProps,
+      ...others
+    } = this.props
 
     const styleCard = [styles.card, isDisabled && styles.card.isDisabled]
 
     return (
-      <StyleRoot style={styleCard}>
+      <StyleRoot style={styleCard} {...others}>
         <Marger bottom={viewportIsSOrLess ? 0 : 5}>
           <Grid style={styles.card.addPadding} disabled={isDisabled}>
             <GridCol col-m="7">{this.renderDescription()}</GridCol>
@@ -218,7 +249,14 @@ class RewardCardComponent extends Component {
   }
 
   renderButton() {
-    const { button, isCompleted, viewportIsSOrLess } = this.props
+    const {
+      button,
+      isCompleted,
+      viewportIsSOrLess,
+      buttonOnMouseEnter,
+      buttonOnMouseLeave,
+      buttonOnClick,
+    } = this.props
 
     if (!button || isCompleted) return
 
@@ -229,6 +267,9 @@ class RewardCardComponent extends Component {
           modifier="helium"
           aria-label={button}
           style={styles.button}
+          onMouseEnter={buttonOnMouseEnter}
+          onMouseLeave={buttonOnMouseLeave}
+          onClick={buttonOnClick}
         >
           {button}
         </Button>